wd-otel-core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wd-otel-core"
version = "0.1.0"
description = "WD OpenTelemetry SDK — core config, setup, and helpers"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-exporter-prometheus>=0.45b0",
    "opentelemetry-instrumentation-logging>=0.45b0",
    "prometheus-client>=0.20.0",
    "python-logging-loki>=0.3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
